requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
python-dotenv
psycopg2-binary
werkzeug
